blog_rest/requirements.txt
amqp==2.5.2
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
asn1crypto==0.24.0
billiard==3.6.3.0
Brlapi==0.6.6
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==2.2.11
django-adminrestrict==2.0.2
django-autoslug==1.9.6
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-countries==6.0
django-extensions==2.2.8
django-filter==2.2.0
django-ipware==2.1.0
django-modeladmin-reorder==0.3.1
django-multiselectfield==0.1.12
django-multitenant==2.2.0
django-postgres-metrics==0.6.2
django-rest-swagger==2.2.0
django-structlog==1.5.0
django-templated-mail==1.1.1
django-timezone-field==4.0
djangorestframework==3.11.0
djoser==2.0.3
drf-friendly-errors==0.13
drf-yasg==1.17.1
dry-rest-permissions==0.1.10
filelock==3.0.12
gunicorn==20.0.4
httplib2==0.9.2
idna==2.9
importlib-metadata==1.7.0
importlib-resources==3.0.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.8
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.1.1
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
openapi-codec==1.3.2
packaging==20.3
pexpect==4.2.1
phonenumbers==8.11.5
Pillow==7.0.0
pipenv==2020.6.2
protobuf==3.0.0
psycopg2==2.8.4
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-crontab==2.4.0
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
SecretStorage==2.3.1
sentry-sdk==0.14.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
ssh-import-id==5.7
structlog==20.1.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.3
vine==1.3.0
virtualenv==20.0.25
virtualenv-clone==0.5.4
wadllib==1.3.2
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
